debugging floating point errors
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@3602 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netimport/visum/NIVisumParser_Lanes.cpp
@@ -187,12 +187,12 @@ NIVisumParser_Lanes::myDependentReport()
             if (edgeID.substr(edgeID.length()-node->getID().length()-1)=="_" + node->getID()) {
                 edgeID = edgeID.substr(0, edgeID.find('_'));
             }
-            NBNode *rn = new NBNode(edgeID + "_" +  toString(length) + "_" + node->getID(), p);
+            NBNode *rn = new NBNode(edgeID + "_" +  toString((size_t) length) + "_" + node->getID(), p);
             if (!myNodeCont.insert(rn)) {
                 MsgHandler::getErrorInstance()->inform("Ups - could not insert node!");
                 throw ProcessError();
             }
-            string nid = edgeID + "_" +  toString(length) + "_" + node->getID();
+            string nid = edgeID + "_" +  toString((size_t) length) + "_" + node->getID();
             myEdgeCont.splitAt(myDistrictCont, edge, useLength/*edge->getLength()-length*/, rn,
                                edge->getID(), nid,
                                edge->getNoLanes()+0, edge->getNoLanes()+1);
